Correction to subtract_line when applied to non-square image
For non-square image and line subtraction using the median, the tiles were created with the wrong shape. For an image shape im_shape, the median over axis 0 col_median produces an array of length im_shape[1]. Therefore, col_median needs to be repeated im_shape[0] (and not im_shape[1]) times to reproduce an image of shape im_shape. 
Same problem with row_median.
This was of course not a problem for square images.